Add proper squares

+++ monopolish/random.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 
 // Return a random number from 0 to 6.
-const uint8_t Random()
+const uint32_t Random()
 {
-	return static_cast<int>(static_cast<double>(rand()) / (static_cast<int64_t>(RAND_MAX + 1)) * 6.0f + 1);
+	return static_cast<uint32_t>(static_cast<double>(rand()) / (static_cast<int64_t>(RAND_MAX + 1)) * 6.0f + 1);
 }
